refactored structure of frontend, added created games list in create view
